fix(responses): support the include query parameter on store retrieval
Extract the include functionality from conversations/ into a shared
openai::include module and wire GET /v1/responses/{id} and its
input_items through it, replacing the hardcoded rejection.

Fixes #685
