+++ apis/src/openai/conversations/contracts.rs
@@ -20,91 +20,6 @@ pub(super) const MAX_ITEMS_PER_REQUEST: usize = 20;
 
 /// Optional response fields supported by Conversation item endpoints.
 ///
-/// The spellings match OpenAI's `IncludeEnum` exactly. Runtime query parsing
-/// and generated `OpenAPI` both consume this enum.
-#[derive(Clone, Copy, Debug, Deserialize, Eq, PartialEq, Serialize, ToSchema)]
-pub(super) enum IncludeField {
-    /// Include file search result payloads.
-    #[serde(rename = "file_search_call.results")]
-    #[schema(rename = "file_search_call.results")]
-    FileSearchCallResults,
-    /// Include web search result payloads.
-    #[serde(rename = "web_search_call.results")]
-    #[schema(rename = "web_search_call.results")]
-    WebSearchCallResults,
-    /// Include the sources used by web search actions.
-    #[serde(rename = "web_search_call.action.sources")]
-    #[schema(rename = "web_search_call.action.sources")]
-    WebSearchCallActionSources,
-    /// Include image URLs in message input-image parts.
-    #[serde(rename = "message.input_image.image_url")]
-    #[schema(rename = "message.input_image.image_url")]
-    MessageInputImageImageUrl,
-    /// Include image URLs in computer-call outputs.
-    #[serde(rename = "computer_call_output.output.image_url")]
-    #[schema(rename = "computer_call_output.output.image_url")]
-    ComputerCallOutputImageUrl,
-    /// Include code-interpreter output payloads.
-    #[serde(rename = "code_interpreter_call.outputs")]
-    #[schema(rename = "code_interpreter_call.outputs")]
-    CodeInterpreterCallOutputs,
-    /// Include encrypted reasoning content.
-    #[serde(rename = "reasoning.encrypted_content")]
-    #[schema(rename = "reasoning.encrypted_content")]
-    ReasoningEncryptedContent,
-    /// Include token log probabilities in message output-text parts.
-    #[serde(rename = "message.output_text.logprobs")]
-    #[schema(rename = "message.output_text.logprobs")]
-    MessageOutputTextLogprobs,
-}
-
-impl IncludeField {
-    /// Parse one decoded query value using the official enum spelling.
-    pub(super) fn parse(value: &str) -> Option<Self> {
-        match value {
-            "file_search_call.results" => Some(Self::FileSearchCallResults),
-            "web_search_call.results" => Some(Self::WebSearchCallResults),
-            "web_search_call.action.sources" => Some(Self::WebSearchCallActionSources),
-            "message.input_image.image_url" => Some(Self::MessageInputImageImageUrl),
-            "computer_call_output.output.image_url" => Some(Self::ComputerCallOutputImageUrl),
-            "code_interpreter_call.outputs" => Some(Self::CodeInterpreterCallOutputs),
-            "reasoning.encrypted_content" => Some(Self::ReasoningEncryptedContent),
-            "message.output_text.logprobs" => Some(Self::MessageOutputTextLogprobs),
-            _ => None,
-        }
-    }
-
-    /// Return this field's bit in the compact runtime include set.
-    const fn bit(self) -> u8 {
-        match self {
-            Self::FileSearchCallResults => 1 << 0,
-            Self::WebSearchCallResults => 1 << 1,
-            Self::WebSearchCallActionSources => 1 << 2,
-            Self::MessageInputImageImageUrl => 1 << 3,
-            Self::ComputerCallOutputImageUrl => 1 << 4,
-            Self::CodeInterpreterCallOutputs => 1 << 5,
-            Self::ReasoningEncryptedContent => 1 << 6,
-            Self::MessageOutputTextLogprobs => 1 << 7,
-        }
-    }
-}
-
-/// Set of requested optional item fields.
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-pub(super) struct IncludeFields(u8);
-
-impl IncludeFields {
-    /// Add one requested field.
-    pub(super) fn insert(&mut self, field: IncludeField) {
-        self.0 |= field.bit();
-    }
-
-    /// Return whether a field was requested.
-    pub(super) const fn contains(self, field: IncludeField) -> bool {
-        self.0 & field.bit() != 0
-    }
-}
-
 /// Request body accepted by `POST /conversations`.
 #[derive(Debug, Default, Deserialize, ToSchema)]
 pub(super) struct CreateConversationRequest {
