+++ apis/src/openai/conversations/handlers.rs
@@ -14,10 +14,11 @@ use serde::{
 use serde_json::{Map, Value};
 use tracing::debug;
 
+use crate::openai::include::{IncludeFields, parse_include, project_item, decode_query_component_strict};
 use super::{
     contracts::{
         ConversationItem, ConversationItemList, ConversationResource, CreateConversationItemsRequest,
-        CreateConversationRequest, DeletedConversationResource, IncludeField, IncludeFields, ItemOrder,
+        CreateConversationRequest, DeletedConversationResource, ItemOrder,
         MAX_ITEMS_PER_REQUEST, Metadata, UpdateConversationRequest,
     },
     validate::{MetadataError, validate_metadata},
@@ -260,7 +261,7 @@ pub(super) async fn handle_create_items(
         Ok(v) => v,
         Err(msg) => return Ok(FilterAction::Reject(invalid_input_response(&msg)?)),
     };
-    let includes = match parse_include_fields(ctx.request.uri.query()) {
+    let includes = match parse_include(ctx.request.uri.query()) {
         Ok(includes) => includes,
         Err(msg) => return Ok(FilterAction::Reject(invalid_input_response(&msg)?)),
     };
@@ -336,7 +337,7 @@ pub(super) async fn handle_list_items(
     conversation_id: &str,
 ) -> Result<FilterAction, FilterError> {
     let tenant_id = ctx.get_metadata(TENANT_METADATA_KEY).unwrap_or(DEFAULT_TENANT_ID);
-    let includes = match parse_include_fields(ctx.request.uri.query()) {
+    let includes = match parse_include(ctx.request.uri.query()) {
         Ok(includes) => includes,
         Err(msg) => return Ok(FilterAction::Reject(invalid_input_response(&msg)?)),
     };
@@ -385,7 +386,7 @@ pub(super) async fn handle_get_item(
     item_id: &str,
 ) -> Result<FilterAction, FilterError> {
     let tenant_id = ctx.get_metadata(TENANT_METADATA_KEY).unwrap_or(DEFAULT_TENANT_ID);
-    let includes = match parse_include_fields(ctx.request.uri.query()) {
+    let includes = match parse_include(ctx.request.uri.query()) {
         Ok(includes) => includes,
         Err(msg) => return Ok(FilterAction::Reject(invalid_input_response(&msg)?)),
     };
@@ -397,7 +398,7 @@ pub(super) async fn handle_get_item(
     match store.get_conversation_item(tenant_id, conversation_id, item_id).await {
         Ok(Some(record)) => {
             let mut item_data = record.item_data;
-            project_conversation_item(&mut item_data, includes);
+            project_item(&mut item_data, includes);
             let item = ConversationItem::from_value(item_data);
             Ok(FilterAction::Reject(json_response(200, &item)?))
         },
@@ -648,162 +649,11 @@ fn conversation_items_response(
             last_id = record.item_id;
         }
         let mut item_data = record.item_data;
-        project_conversation_item(&mut item_data, includes);
+        project_item(&mut item_data, includes);
         data.push(ConversationItem::from_value(item_data));
     }
 
     ConversationItemList::new(data, has_more, first_id, last_id)
-}
-
-/// Remove optional fields that were not requested through `include`.
-///
-/// Projection changes only the response-owned value after the complete item
-/// representation has crossed the storage boundary.
-fn project_conversation_item(item: &mut Value, includes: IncludeFields) {
-    let Some(object) = item.as_object_mut() else {
-        return;
-    };
-    match projection_kind(object) {
-        ProjectionKind::Reasoning => remove_unless_included(
-            object,
-            "encrypted_content",
-            includes.contains(IncludeField::ReasoningEncryptedContent),
-        ),
-        ProjectionKind::FileSearch => remove_unless_included(
-            object,
-            "results",
-            includes.contains(IncludeField::FileSearchCallResults),
-        ),
-        ProjectionKind::WebSearch => project_web_search_fields(object, includes),
-        ProjectionKind::CodeInterpreter => remove_unless_included(
-            object,
-            "outputs",
-            includes.contains(IncludeField::CodeInterpreterCallOutputs),
-        ),
-        ProjectionKind::ComputerOutput => project_computer_output_fields(object, includes),
-        ProjectionKind::Message => project_message_fields(object, includes),
-        ProjectionKind::Other => {},
-    }
-}
-
-/// Item variants with fields controlled by `include`.
-#[derive(Clone, Copy)]
-enum ProjectionKind {
-    /// Reasoning item with optional encrypted content.
-    Reasoning,
-    /// File-search call with optional results.
-    FileSearch,
-    /// Web-search call with optional results and sources.
-    WebSearch,
-    /// Code-interpreter call with optional outputs.
-    CodeInterpreter,
-    /// Computer-call output with an optional image URL.
-    ComputerOutput,
-    /// Message with optional fields in typed content parts.
-    Message,
-    /// Item without any fields controlled by `include`.
-    Other,
-}
-
-/// Classify an item without retaining a borrow into the mutable object.
-fn projection_kind(object: &Map<String, Value>) -> ProjectionKind {
-    match object.get("type").and_then(Value::as_str) {
-        Some("reasoning") => ProjectionKind::Reasoning,
-        Some("file_search_call") => ProjectionKind::FileSearch,
-        Some("web_search_call") => ProjectionKind::WebSearch,
-        Some("code_interpreter_call") => ProjectionKind::CodeInterpreter,
-        Some("computer_call_output") => ProjectionKind::ComputerOutput,
-        Some("message") => ProjectionKind::Message,
-        _ => ProjectionKind::Other,
-    }
-}
-
-/// Remove one top-level field unless it was explicitly requested.
-fn remove_unless_included(object: &mut Map<String, Value>, field: &str, included: bool) {
-    if !included {
-        object.remove(field);
-    }
-}
-
-/// Project web-search fields controlled by independent include values.
-fn project_web_search_fields(object: &mut Map<String, Value>, includes: IncludeFields) {
-    remove_unless_included(object, "results", includes.contains(IncludeField::WebSearchCallResults));
-    if !includes.contains(IncludeField::WebSearchCallActionSources)
-        && let Some(action) = object.get_mut("action").and_then(Value::as_object_mut)
-    {
-        action.remove("sources");
-    }
-}
-
-/// Project the nested image URL from a computer-call output.
-fn project_computer_output_fields(object: &mut Map<String, Value>, includes: IncludeFields) {
-    if !includes.contains(IncludeField::ComputerCallOutputImageUrl)
-        && let Some(output) = object.get_mut("output").and_then(Value::as_object_mut)
-    {
-        output.remove("image_url");
-    }
-}
-
-/// Project optional fields from typed message content parts.
-fn project_message_fields(object: &mut Map<String, Value>, includes: IncludeFields) {
-    let Some(content) = object.get_mut("content").and_then(Value::as_array_mut) else {
-        return;
-    };
-    for part in content {
-        let Some(part) = part.as_object_mut() else {
-            continue;
-        };
-        if part.get("type").and_then(Value::as_str) == Some("input_image")
-            && !includes.contains(IncludeField::MessageInputImageImageUrl)
-        {
-            part.remove("image_url");
-        } else if part.get("type").and_then(Value::as_str) == Some("output_text")
-            && !includes.contains(IncludeField::MessageOutputTextLogprobs)
-        {
-            part.remove("logprobs");
-        }
-    }
-}
-
-/// Parse both official SDK encodings for the array-valued `include` query:
-/// repeated `include=value` pairs and bracketed `include[]=value` pairs.
-fn parse_include_fields(query: Option<&str>) -> Result<IncludeFields, String> {
-    let Some(query) = query else {
-        return Ok(IncludeFields::default());
-    };
-
-    let mut includes = IncludeFields::default();
-    for pair in query.split('&') {
-        let Some((raw_key, raw_value)) = pair.split_once('=') else {
-            let key = decode_query_component_strict(pair)?;
-            if matches!(key.as_ref(), "include" | "include[]") {
-                return Err("'include' query parameter requires a value".to_owned());
-            }
-            continue;
-        };
-        let key = decode_query_component_strict(raw_key)?;
-        if !matches!(key.as_ref(), "include" | "include[]") {
-            continue;
-        }
-        let value = decode_query_component_strict(raw_value)?;
-        let field = IncludeField::parse(&value).ok_or_else(|| format!("unsupported include value: '{value}'"))?;
-        includes.insert(field);
-    }
-    Ok(includes)
-}
-
-/// Strictly decode one query component, including form-style `+` spaces.
-fn decode_query_component_strict(value: &str) -> Result<Cow<'_, str>, String> {
-    if value.contains('+') {
-        let normalized = value.replace('+', " ");
-        return percent_decode_str(&normalized)
-            .decode_utf8()
-            .map(|decoded| Cow::Owned(decoded.into_owned()))
-            .map_err(|e| format!("query parameter must be valid UTF-8: {e}"));
-    }
-    percent_decode_str(value)
-        .decode_utf8()
-        .map_err(|e| format!("query parameter must be valid UTF-8: {e}"))
 }
 
 /// Parse and validate cursor-based pagination parameters from a query string.
@@ -1340,161 +1190,6 @@ mod tests {
             err.contains("Duplicate query parameter: 'limit'"),
             "encoded duplicate key should be rejected: {err}"
         );
-    }
-
-    // -------------------------------------------------------------------------
-    // include parsing and projection
-    // -------------------------------------------------------------------------
-
-    #[test]
-    fn parse_include_fields_supports_python_and_node_sdk_encodings() {
-        let includes = parse_include_fields(Some(
-            "include=reasoning.encrypted_content&include%5B%5D=message.output_text.logprobs",
-        ))
-        .unwrap();
-
-        assert!(
-            includes.contains(IncludeField::ReasoningEncryptedContent),
-            "repeated-key encoding should parse reasoning encrypted content"
-        );
-        assert!(
-            includes.contains(IncludeField::MessageOutputTextLogprobs),
-            "bracket encoding should parse output-text log probabilities"
-        );
-        assert!(
-            !includes.contains(IncludeField::FileSearchCallResults),
-            "unrequested include values must remain absent"
-        );
-    }
-
-    #[test]
-    fn parse_include_fields_rejects_unknown_or_malformed_values() {
-        let unknown = parse_include_fields(Some("include=future.secret_field")).unwrap_err();
-        assert!(
-            unknown.contains("unsupported include value"),
-            "unknown values should produce an unsupported-value diagnostic: {unknown}"
-        );
-
-        let missing = parse_include_fields(Some("include")).unwrap_err();
-        assert!(
-            missing.contains("requires a value"),
-            "missing include values should identify the required value: {missing}"
-        );
-
-        let invalid_utf8 = parse_include_fields(Some("include=%FF")).unwrap_err();
-        assert!(
-            invalid_utf8.contains("valid UTF-8"),
-            "invalid encoding should identify the UTF-8 requirement: {invalid_utf8}"
-        );
-    }
-
-    #[test]
-    #[expect(clippy::too_many_lines, reason = "one fixture covers every include projection path")]
-    fn projection_removes_every_unrequested_include_gated_field() {
-        let mut items = vec![
-            serde_json::json!({
-                "type": "reasoning",
-                "encrypted_content": "secret",
-                "summary": []
-            }),
-            serde_json::json!({
-                "type": "file_search_call",
-                "results": [{"file_id": "file_1"}],
-                "status": "completed"
-            }),
-            serde_json::json!({
-                "type": "web_search_call",
-                "results": [{"url": "https://example.com"}],
-                "action": {
-                    "type": "search",
-                    "sources": [{"type": "url", "url": "https://example.com"}]
-                }
-            }),
-            serde_json::json!({
-                "type": "code_interpreter_call",
-                "outputs": [{"type": "logs", "logs": "done"}],
-                "status": "completed"
-            }),
-            serde_json::json!({
-                "type": "computer_call_output",
-                "output": {"type": "computer_screenshot", "image_url": "data:image/png;base64,AA=="}
-            }),
-            serde_json::json!({
-                "type": "message",
-                "content": [
-                    {"type": "input_image", "image_url": "https://example.com/image.png", "detail": "auto"},
-                    {"type": "output_text", "text": "answer", "annotations": [], "logprobs": []},
-                    {"type": "input_text", "text": "keep me"}
-                ]
-            }),
-        ];
-
-        for item in &mut items {
-            project_conversation_item(item, IncludeFields::default());
-        }
-
-        assert!(
-            items[0].get("encrypted_content").is_none(),
-            "reasoning encrypted content should be omitted"
-        );
-        assert!(
-            items[1].get("results").is_none(),
-            "file-search results should be omitted"
-        );
-        assert!(
-            items[2].get("results").is_none(),
-            "web-search results should be omitted"
-        );
-        assert!(
-            items[2]["action"].get("sources").is_none(),
-            "web-search action sources should be omitted"
-        );
-        assert!(
-            items[3].get("outputs").is_none(),
-            "code-interpreter outputs should be omitted"
-        );
-        assert!(
-            items[4]["output"].get("image_url").is_none(),
-            "computer-output image URLs should be omitted"
-        );
-        assert!(
-            items[5]["content"][0].get("image_url").is_none(),
-            "message input-image URLs should be omitted"
-        );
-        assert!(
-            items[5]["content"][1].get("logprobs").is_none(),
-            "message output-text log probabilities should be omitted"
-        );
-        assert_eq!(items[5]["content"][2]["text"], "keep me");
-    }
-
-    #[test]
-    fn projection_preserves_every_requested_include_gated_field() {
-        let mut includes = IncludeFields::default();
-        for field in [
-            IncludeField::FileSearchCallResults,
-            IncludeField::WebSearchCallResults,
-            IncludeField::WebSearchCallActionSources,
-            IncludeField::MessageInputImageImageUrl,
-            IncludeField::ComputerCallOutputImageUrl,
-            IncludeField::CodeInterpreterCallOutputs,
-            IncludeField::ReasoningEncryptedContent,
-            IncludeField::MessageOutputTextLogprobs,
-        ] {
-            includes.insert(field);
-        }
-        let original = serde_json::json!({
-            "type": "message",
-            "content": [
-                {"type": "input_image", "image_url": "https://example.com/image.png"},
-                {"type": "output_text", "logprobs": [{"token": "x"}]}
-            ]
-        });
-        let mut projected = original.clone();
-
-        project_conversation_item(&mut projected, includes);
-
-        assert_eq!(projected, original);
     }
 
     // -------------------------------------------------------------------------
