+++ apis/src/openai/responses/store/tests.rs
@@ -21,7 +21,7 @@ use super::{
     list_input_items,
 };
 use crate::{
-    openai::responses::state::ResponsesState,
+    openai::{include::IncludeFields, responses::state::ResponsesState},
     store::{ResponseRecord, ResponseStore as _, ResponseStoreRegistry, SqliteResponseStore},
 };
 
@@ -3725,24 +3725,6 @@ fn parse_query_params_unknown_order_rejected() {
 }
 
 #[test]
-fn parse_query_params_include_bracket_rejected() {
-    let err = super::filter::parse_query_params(Some("include[]=reasoning.encrypted_content")).unwrap_err();
-    assert!(
-        err.contains("not supported"),
-        "should reject include[] parameter: {err}"
-    );
-}
-
-#[test]
-fn parse_query_params_include_bare_rejected() {
-    let err = super::filter::parse_query_params(Some("include=reasoning.encrypted_content")).unwrap_err();
-    assert!(
-        err.contains("not supported"),
-        "should reject bare include parameter: {err}"
-    );
-}
-
-#[test]
 fn parse_query_params_unknown_parameter_rejected() {
     let err = super::filter::parse_query_params(Some("foo=bar")).unwrap_err();
     assert!(
@@ -3969,7 +3951,9 @@ fn make_record_with_input(input: serde_json::Value) -> ResponseRecord {
 #[test]
 fn list_input_items_scalar_input_normalized_to_message_item() {
     let record = make_record_with_input(json!("Hello"));
-    let page = list_input_items(&record, &ListParams::default()).unwrap();
+    let params = ListParams::default();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert_eq!(page.data.len(), 1, "string input should produce one message item");
     assert_eq!(
         page.data[0],
@@ -3987,7 +3971,9 @@ fn list_input_items_scalar_input_normalized_to_message_item() {
 #[test]
 fn list_input_items_null_input_returns_empty_page() {
     let record = make_record_with_input(json!(null));
-    let page = list_input_items(&record, &ListParams::default()).unwrap();
+    let params = ListParams::default();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert!(page.data.is_empty(), "null input should yield no items");
     assert!(!page.has_more);
     assert!(page.next_cursor.is_none());
@@ -3996,7 +3982,9 @@ fn list_input_items_null_input_returns_empty_page() {
 #[test]
 fn list_input_items_empty_array_returns_empty_page() {
     let record = make_record_with_input(json!([]));
-    let page = list_input_items(&record, &ListParams::default()).unwrap();
+    let params = ListParams::default();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert!(page.data.is_empty(), "empty array input should yield empty page");
     assert!(!page.has_more);
     assert!(page.next_cursor.is_none());
@@ -4013,7 +4001,8 @@ fn list_input_items_ascending_preserves_natural_order() {
         order: Order::Ascending,
         ..Default::default()
     };
-    let page = list_input_items(&record, &params).unwrap();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert_eq!(page.data[0]["id"], "a");
     assert_eq!(page.data[1]["id"], "b");
     assert_eq!(page.data[2]["id"], "c");
@@ -4030,7 +4019,8 @@ fn list_input_items_descending_reverses_order() {
         order: Order::Descending,
         ..Default::default()
     };
-    let page = list_input_items(&record, &params).unwrap();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert_eq!(page.data[0]["id"], "c");
     assert_eq!(page.data[1]["id"], "b");
     assert_eq!(page.data[2]["id"], "a");
@@ -4048,7 +4038,8 @@ fn list_input_items_limit_zero_clamped_to_one() {
         order: Order::Ascending,
         ..Default::default()
     };
-    let page = list_input_items(&record, &params).unwrap();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert_eq!(page.data.len(), 1, "limit=0 should be clamped to 1");
     assert!(page.has_more);
 }
@@ -4062,7 +4053,8 @@ fn list_input_items_limit_above_max_clamped() {
         order: Order::Ascending,
         ..Default::default()
     };
-    let page = list_input_items(&record, &params).unwrap();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert_eq!(
         page.data.len(),
         5,
@@ -4082,7 +4074,8 @@ fn list_input_items_cursor_after_last_returns_empty_page() {
         order: Order::Ascending,
         ..Default::default()
     };
-    let page = list_input_items(&record, &params).unwrap();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert!(page.data.is_empty(), "cursor after last item should yield empty page");
     assert!(!page.has_more);
 }
@@ -4099,7 +4092,8 @@ fn list_input_items_cursor_after_first_skips_it() {
         order: Order::Ascending,
         ..Default::default()
     };
-    let page = list_input_items(&record, &params).unwrap();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert_eq!(page.data.len(), 2);
     assert_eq!(page.data[0]["id"], "b");
     assert_eq!(page.data[1]["id"], "c");
@@ -4117,7 +4111,8 @@ fn list_input_items_numeric_cursor_fallback() {
         order: Order::Ascending,
         ..Default::default()
     };
-    let page = list_input_items(&record, &params).unwrap();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert_eq!(page.data.len(), 2, "numeric cursor 1 should skip first item");
     assert_eq!(page.data[0]["val"], 2);
     assert_eq!(page.data[1]["val"], 3);
@@ -4134,7 +4129,8 @@ fn list_input_items_invalid_cursor_returns_error() {
         order: Order::Ascending,
         ..Default::default()
     };
-    let result = list_input_items(&record, &params);
+    let includes = IncludeFields::default();
+    let result = list_input_items(&record, &params, includes);
     assert!(result.is_err(), "non-numeric cursor not matching any ID should error");
 }
 
@@ -4151,7 +4147,8 @@ fn list_input_items_next_cursor_uses_item_id() {
         order: Order::Ascending,
         ..Default::default()
     };
-    let page = list_input_items(&record, &params).unwrap();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert!(page.has_more);
     assert_eq!(
         page.next_cursor.as_deref(),
@@ -4168,7 +4165,9 @@ fn list_input_items_assigns_synthetic_ids_to_id_less_array_items() {
     ]));
     // Default order is descending, but synthetic IDs must stay tied to
     // each item's original stored position, not its display position.
-    let page = list_input_items(&record, &ListParams::default()).unwrap();
+    let params = ListParams::default();
+    let includes = IncludeFields::default();    
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert_eq!(
         page.data[0]["id"], "msg_resp_test_input_1",
         "the assistant message (original index 1) should surface first in descending order"
@@ -4191,7 +4190,8 @@ fn list_input_items_mixed_existing_and_synthetic_ids_paginate_correctly() {
         ..Default::default()
     };
 
-    let page1 = list_input_items(&record, &base_params).unwrap();
+    let includes = IncludeFields::default();
+    let page1 = list_input_items(&record, &base_params, includes).unwrap();
     assert_eq!(
         page1.data[0]["id"], "a",
         "the item's existing id must be preserved, not overwritten"
@@ -4210,6 +4210,7 @@ fn list_input_items_mixed_existing_and_synthetic_ids_paginate_correctly() {
             cursor: cursor_after_a,
             ..base_params.clone()
         },
+        includes,
     )
     .unwrap();
     assert_eq!(
@@ -4228,6 +4229,7 @@ fn list_input_items_mixed_existing_and_synthetic_ids_paginate_correctly() {
             cursor: Some("msg_resp_test_input_1".to_owned()),
             ..base_params
         },
+        includes,
     )
     .unwrap();
     assert!(
@@ -4250,7 +4252,8 @@ fn list_input_items_next_cursor_uses_synthetic_id_when_input_lacks_ids() {
         order: Order::Ascending,
         ..Default::default()
     };
-    let page = list_input_items(&record, &params).unwrap();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert!(page.has_more);
     assert_eq!(
         page.next_cursor.as_deref(),
@@ -4272,7 +4275,8 @@ fn list_input_items_numeric_cursor_still_works_as_fallback_for_synthetic_ids() {
         order: Order::Ascending,
         ..Default::default()
     };
-    let page = list_input_items(&record, &params).unwrap();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert_eq!(
         page.data.len(),
         2,
@@ -4290,7 +4294,8 @@ fn list_input_items_no_next_cursor_when_no_more() {
         order: Order::Ascending,
         ..Default::default()
     };
-    let page = list_input_items(&record, &params).unwrap();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert!(!page.has_more);
     assert!(
         page.next_cursor.is_none(),
@@ -4309,7 +4314,9 @@ fn list_input_items_default_limit_is_default_page_limit() {
 #[test]
 fn list_input_items_object_input_wrapped_as_single_item() {
     let record = make_record_with_input(json!({"type": "message", "role": "user", "content": "hi"}));
-    let page = list_input_items(&record, &ListParams::default()).unwrap();
+    let params = ListParams::default();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert_eq!(page.data.len(), 1);
     assert_eq!(page.data[0]["type"], "message");
 }
@@ -4325,9 +4332,10 @@ fn list_input_items_descending_cursor_operates_on_reversed_order() {
     let params = ListParams {
         cursor: Some("c".to_owned()),
         limit: 2,
-        order: Order::Descending,
+        order: Order::Descending,   
     };
-    let page = list_input_items(&record, &params).unwrap();
+    let includes = IncludeFields::default();
+    let page = list_input_items(&record, &params, includes).unwrap();
     assert_eq!(page.data.len(), 2, "should return items after cursor in reversed order");
     assert_eq!(page.data[0]["id"], "b");
     assert_eq!(page.data[1]["id"], "a");
